+++ web/src/components/ui/index.ts
@@ -680,3 +680,4 @@ export * from './chart-line-rmi-mid-cross';
 export * from './chart-line-fisher-overbought-cross';
 export * from './chart-line-fisher-oversold-cross';
 export * from './chart-line-rmi-overbought-cross';
+export * from './chart-line-rmi-oversold-cross';